Extract TLS Version from cookie and send to GA
The TLS Version supported by the browser will be set as a cookie by the
CDN and so will be available when viewing the page.  We extract the value
from the cookie and send it as the value for custom dimension 16 to
google analytics as part of the initial page view tracking.  Doing it
here has caused some churn in the tests (it's now the 4th arg instead
of the trackPageView message), but it means we don't have to send a
separate GA request so it shouldn't impact performance for the users.